Make the test for Web Notifications correctly detect the standard version and the moz prefixed version.
The standard at http://www.w3.org/TR/notifications/ (which is what
data.js cites as the source for the feature) describes the mechanism for
creating notifications as "new Notification()" (no final s).  The older
WebKit implementation, on the other hand, creates notifications with
window.webkitNotifications.createNotification() (with an s on
Notifications).  The early Gecko implementation (available on mobile but
not desktop) uses window.navigator.mozNotification.createNotification()
(no s on notifications, and on window.navigator).

So this patch removes the s from the test for everything except for
window.webkitNotifications, and tests for mozNotification on
window.navigator rather than on window.  I don't see any evidence of an
Opera or Microsoft prefixed implementation based on searching Google for
documentation.