Reduce points for old and prefixed implementations
* encourage vendors to fix their old implementations instead of just
chasing the new and shiny. Gaining extra points for updating could
really move the needle and help web compat/interop, and make updating
more rewarding to vendors, and encourage management buy off to spend
the time.
* prefixes are the single biggest issue affecting interop, along with
old vendor specific implementations that are not updated to a newer
spec that other browsers may support.
* it is a huge burden on non-webkit browsers to have to reverse
engineer and implement webkit quirks and versions of APIs, which may be
very different to the latest spec.
* CSS3test will update to similarly reduce points for prefixes,
prompted by Mozilla, so this would follow suite.